Change tests msg_limit counter from usize to u32 and handle when receiver receives an invalid string

+++ src/tests.rs
@@ -11,7 +11,7 @@ pub enum Action {
         message: String
     },
     Die {
-        after_n_messages: usize
+        after_n_messages: u32
     }
 }
